+++ templates/consts.js
@@ -1,4 +1,9 @@
-export const supportedTools = ["react-router", "firebase", "netlify-functions"]
+export const supportedTools = [
+  "react-router",
+  "firebase",
+  "netlify-functions",
+  "back4app",
+]
 export const comingSoonTools = [
   "zustand",
   "neobrutalism",