+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.2.0] - 2026-01-13
+
+### Added
+- Back4App support with `a10ha add back4app` command
+- Automatic Back4App initialization in App file
+
 ## [1.1.1] - 2026-01-13
 
 ### Changed